Extract particle methods to a different file
